backend/requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
# django-cors-headers==3.8.0 #pour la securite
# channels==4.0.0
# daphne==4.0.0
Pillow==9.0
djangorestframework==3.15.1
requests==2.31.0